@@ -1,23 +1,34 @@
 [package]
-name = "cdt-rs"
+name = "causal-dynamical-triangulations"
 version = "0.1.0"
 authors = ["Adam Getchell <adam@adamgetchell.org>"]
-edition = "2021"
-description = "Causal Dynamical Triangulations using Constrained Delaunay Triangulations in Rust"
+edition = "2024"
+description = "Causal Dynamical Triangulations"
 readme = "README.md"
+license = "BSD-3-Clause"
+repository = "https://github.com/acgetchell/causal-dynamical-triangulations"
+keywords = ["triangulation", "delaunay", "computational-geometry", "physics", "causal"]
+categories = ["science", "mathematics", "algorithms"]
+
+[[bin]]
+name = "cdt-rs"
+path = "src/main.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 clap = { version = "4.5.20", features = ["derive"] }
 float-ord = "0.3.2"
-rand = "0.8.5"
-spade = { version = "2.12.1", features = ["serde"] }
-assert_cmd = "2.0.16"
+rand = "0.9.2"
+delaunay = "0.5.0"
+assert_cmd = "2.0.17"
 
 
 [dev-dependencies]
-predicates = "3.1.2"
+predicates = "3.1.3"
 
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(kani)'] }
+
+[lints.clippy]
+multiple_crate_versions = "allow"
